+++ salve_ipc/__init__.py
@@ -1,4 +1,7 @@
-from .highlight import Token, generic_tokens  # noqa: F401
 from .ipc import IPC  # noqa: F401
 from .misc import COMMANDS, Request, Response  # noqa: F401
-from .server_functions import is_unicode_letter  # noqa: F401
+from .server_functions import (
+    is_unicode_letter,
+    Token,
+    generic_tokens,
+)  # noqa: F401
