--- salve_ipc/highlight/__init__.py
@@ -1,6 +1,0 @@
-from .highlight import (  # noqa: F401
-    Token,
-    generic_tokens,
-    get_highlights,
-    hidden_chars,
-)
