+++ salve_ipc/ipc.py
@@ -1,14 +1,9 @@
-from json import dumps, loads
-from os import remove, set_blocking
-from pathlib import Path
 from random import randint
-from subprocess import PIPE, Popen
-from tempfile import NamedTemporaryFile, _TemporaryFileWrapper
-from typing import IO
-
+from multiprocessing import Pipe, Process, Queue, freeze_support
+from multiprocessing.connection import Connection
+from .server import Server
 from .misc import (
     COMMANDS,
-    Details,
     Message,
     Notification,
     Ping,
@@ -28,7 +23,7 @@ class IPC:
     """
 
     def __init__(self, id_max: int = 15_000) -> None:
-        self.all_ids: dict[int, str] = {}  # id, tempfile path
+        self.all_ids: list[int] = []
         self.id_max = id_max
         self.current_ids: dict[str, int] = {}
         self.newest_responses: dict[str, Response | None] = {}
@@ -38,103 +33,73 @@ class IPC:
 
         self.files: dict[str, str] = {}
 
-        self.main_server: Popen
+        self.response_queue: Queue[Response] = Queue()
+        self.requests_queue: Queue[Request] = Queue()
+        self.client_end: Connection
+        self.main_server: Process
         self.create_server()
 
     def create_server(self) -> None:
         """Creates the main_server through a subprocess - internal API"""
-        server_file: Path = Path(__file__).parent / "server.py"
-        server = Popen(
-            ["python3", str(server_file)], stdin=PIPE, stdout=PIPE, stderr=PIPE
+        self.client_end, server_end = Pipe()
+        freeze_support()
+        self.main_server = Process(
+            target=Server,
+            args=(server_end, self.response_queue, self.requests_queue),
         )
-        set_blocking(server.stdout.fileno(), False)  # type: ignore
-        set_blocking(server.stdin.fileno(), False)  # type: ignore
-        self.main_server = server
+        self.main_server.start()
 
         files_copy = self.files.copy()
         self.files = {}
         for filename, data in files_copy.items():
             self.update_file(filename, data)
 
-    def check_server_error(self) -> tuple[bool, str]:
-        """Returns whether the server is dead or not and, if it is, returns the error as well for logging - external API"""
-        dead: bool = False if not self.main_server.poll() else True  # type: ignore
-        if dead:
-            err: str = self.main_server.stderr.read().decode()  # type: ignore
-            return (True, err)
-        return (False, "")
-
     def check_server(self) -> None:
         """Checks that the main_server is alive - internal API"""
-        if self.main_server.poll() is not None:
+        if not self.main_server.is_alive():
             self.create_server()
-
-    def get_server_file(self, file: str) -> IO:
-        """Returns the main_server stdin/stdout based on the argument provided ("stdin"/"stdout") - internal API"""
-        self.check_server()
-        if file == "stdout":
-            return self.main_server.stdout  # type: ignore
-        return self.main_server.stdin  # type: ignore
-
-    def write_tmp_file(self, details: Details, tmp_path: str) -> None:
-        with open(tmp_path, "r+") as file:
-            file.write(dumps(details))
 
     def send_message(self, message: Message) -> None:
         """Sends a Message to the main_server as provided by the argument message - internal API"""
-        json_request: str = dumps(message)
 
-        server_stdin = self.get_server_file("stdin")
-        server_stdin.write(f"{json_request}\n".encode())
-        server_stdin.flush()
+        self.requests_queue.put(message)  # type: ignore
 
     def create_message(self, type: str, **kwargs) -> None:
         """Creates a Message based on the args and kwawrgs provided. Highly flexible. - internal API"""
         id = randint(1, self.id_max)  # 0 is reserved for the empty case
-        while id in list(self.all_ids.keys()):
+        while id in self.all_ids:
             id = randint(1, self.id_max)
+        self.all_ids.append(id)
 
-        tmp: _TemporaryFileWrapper[str] = NamedTemporaryFile(
-            prefix="salve_ipc", suffix=".tmp", delete=False, mode="r+"
-        )
-        self.all_ids[id] = tmp.name
         match type:
             case "ping":
-                ping: Ping = {"id": id, "type": "ping", "tmp_file": tmp.name}
+                ping: Ping = {"id": id, "type": "ping"}
                 self.send_message(ping)
             case "request":
                 command = kwargs.get("command", "")
                 self.current_ids[command] = id
-                request_details: Request = {
-                    "command": command,
-                    "file": kwargs.get("file"),
-                    "expected_keywords": kwargs.get("expected_keywords"),
-                    "current_word": kwargs.get("current_word"),
-                    "language": kwargs.get("language"),
-                    "text_range": kwargs.get("text_range"),
-                }  # type: ignore
-                request: Message = {
+                request: Request = {
                     "id": id,
                     "type": type,
-                    "tmp_file": tmp.name,
+                    "command": command,
+                    "file": kwargs.get("file", ""),
+                    "expected_keywords": kwargs.get("expected_keywords", []),
+                    "current_word": kwargs.get("current_word", ""),
+                    "language": kwargs.get("language", ""),
+                    "text_range": kwargs.get("text_range", (1, -1)),
                 }
-                self.write_tmp_file(request_details, tmp.name)
                 self.send_message(request)
             case "notification":
-                notification_details: Notification = {
+                notification: Notification = {
+                    "id": id,
+                    "type": type,
                     "remove": kwargs.get("remove", False),
                     "file": kwargs.get("filename", ""),
                     "contents": kwargs.get("contents", ""),
                 }
-                notification: Message = {
-                    "id": id,
-                    "type": type,
-                    "tmp_file": tmp.name,
-                }
-                self.write_tmp_file(notification_details, tmp.name)
                 self.send_message(notification)
             case _:
-                ping: Ping = {"id": id, "type": "ping", "tmp_file": tmp.name}
+                ping: Ping = {"id": id, "type": "ping"}
                 self.send_message(ping)
 
     def ping(self) -> None:
@@ -181,30 +146,25 @@ class IPC:
 
         self.current_ids[command] = 0
 
-    def parse_line(self, line: str) -> None:
+    def parse_response(self, res: Response) -> None:
         """Parses main_server output line and discards useless responses - internal API"""
-        response_json: Message = loads(line)
-        details: Response = loads(open(response_json["tmp_file"], "r+").read())
-        id = response_json["id"]
-        self.all_ids.pop(id)
-        remove(response_json["tmp_file"])
+        id = res["id"]
+        self.all_ids.remove(id)
 
-        if "command" not in details:
+        if "command" not in res:
             return
 
-        command = details["command"]
+        command = res["command"]
         if id != self.current_ids[command]:
             return
 
         self.current_ids[command] = 0
-        self.newest_responses[command] = details
+        self.newest_responses[command] = res
 
     def check_responses(self) -> None:
         """Checks all main_server output by calling IPC.parse_line() on each response - internal API"""
-        server_stdout: IO = self.get_server_file("stdout")
-
-        for line in server_stdout:  # type: ignore
-            self.parse_line(line)
+        while not self.response_queue.empty():
+            self.parse_response(self.response_queue.get())
 
     def get_response(self, command: str) -> Response | None:
         """Runs IPC.check_responses() and returns the current response of type command if it has been returned - external API"""
@@ -216,8 +176,6 @@ class IPC:
 
         self.check_responses()
         response: Response | None = self.newest_responses[command]
-        if response is None:
-            return None
         self.newest_responses[command] = None
         return response
